+++ packages/gatsby-theme-aoi-blog-components/src/index.tsx
@@ -1,4 +1,5 @@
 export { default as DrawerPageNavigation } from './DrawerPageNavigation';
+export { default as muiComponents } from './muiComponents';
 export { default as PageNavigationContainer } from './PageNavigationContainer';
 export { default as PageNavigationItem } from './PageNavigationItem';
 export { default as Pagination } from './Pagination';