+++ packages/gatsby-theme-aoi-components/src/Article.tsx
@@ -5,6 +5,7 @@ import Container, { ContainerProps } from '@mui/material/Container';
 import Typography, { TypographyProps } from '@mui/material/Typography';
 import MuiLink, { LinkProps } from '@mui/material/Link';
 import Divider from '@mui/material/Divider';
+import { alpha } from '@mui/material/styles';
 
 /**
  * `<Article>`
@@ -50,7 +51,7 @@ export function ArticleTitle({ children, ...props }: Props) {
 
 export function Paragraph({ children, ...props }: Props) {
   return (
-    <Typography variant="body1" paragraph {...props}>
+    <Typography variant="body1" paragraph lineHeight={1.8} {...props}>
       {children}
     </Typography>
   );
@@ -58,7 +59,13 @@ export function Paragraph({ children, ...props }: Props) {
 
 export function SubParagraph({ children, ...props }: Props) {
   return (
-    <Typography variant="body2" color="text.secondary" paragraph {...props}>
+    <Typography
+      variant="body2"
+      color="text.secondary"
+      paragraph
+      lineHeight={1.8}
+      {...props}
+    >
       {children}
     </Typography>
   );
@@ -78,7 +85,7 @@ export function H4({ children, ...props }: Props) {
       variant="body1"
       component="h4"
       mt="2em"
-      gutterBottom
+      mb="1.2em"
       fontWeight="bold"
       {...props}
     >
@@ -94,7 +101,8 @@ export function H5({ children, ...props }: Props) {
       component="h5"
       color="text.secondary"
       fontWeight="bold"
-      gutterBottom
+      mt="2em"
+      mb="1em"
       {...props}
     >
       {children}
@@ -109,7 +117,8 @@ export function H6({ children, ...props }: Props) {
       component="h6"
       color="text.secondary"
       fontWeight="bold"
-      gutterBottom
+      mt="2em"
+      mb="1em"
       {...props}
     >
       {children}
@@ -149,4 +158,35 @@ export function Blockquote({ children, ...props }: Props) {
 
 export function Hr() {
   return <Divider sx={{ my: 8 }} />;
+}
+
+export function InlineCode({ children }: Omit<TypographyProps, 'ref'>) {
+  return (
+    <Typography
+      variant="body2"
+      component="code"
+      fontFamily="'Consolas', 'Monaco', 'Andale Mono', 'Ubuntu Mono', 'monospace'"
+      px={0.5}
+      borderRadius={1}
+      bgcolor={({ palette }) =>
+        alpha(palette.secondary.main, palette.action.selectedOpacity)
+      }
+    >
+      {children}
+    </Typography>
+  );
+}
+
+export function Ul(props: Omit<TypographyProps, 'ref'>) {
+  return <Typography component="ul" py={2} m={0} {...props} />;
+}
+
+export function Ol(props: Omit<TypographyProps, 'ref'>) {
+  return <Typography component="ol" py={2} m={0} {...props} />;
+}
+
+export function Li(props: Omit<TypographyProps, 'ref'>) {
+  return (
+    <Typography variant="body1" component="li" lineHeight={1.8} {...props} />
+  );
 }