+++ packages/gatsby-theme-aoi-components/src/index.tsx
@@ -17,6 +17,10 @@ export {
   Link,
   Blockquote,
   Hr,
+  InlineCode,
+  Ul,
+  Ol,
+  Li,
 } from './Article';
 export { default as BottomNavItem } from './BottomNavItem';
 export { default as ExternalLink, ExternalLinkProps } from './ExternalLink';