+++ starters/gatsby-starter-aoi-blog/src/@cieloazul310/gatsby-theme-aoi-blog-templates/icons/index.tsx
@@ -1,0 +1,2 @@
+// `AuthorIcon` must be included
+export * from '../../../icons';